+++ examples/FasterRCNN/model_cascade.py
@@ -88,7 +88,7 @@ class CascadeRCNNHead(object):
                 labels_per_box = tf.gather(self.gt_labels, best_iou_ind)
                 fg_mask = max_iou_per_box >= iou_threshold
                 fg_inds_wrt_gt = tf.boolean_mask(best_iou_ind, fg_mask)
-                labels_per_box = tf.stop_gradient(labels_per_box * tf.to_int64(fg_mask))
+                labels_per_box = tf.stop_gradient(labels_per_box * tf.cast(fg_mask, tf.int64))
                 return BoxProposals(boxes, labels_per_box, fg_inds_wrt_gt)
         else:
             return BoxProposals(boxes)
